Backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.44
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-nmap==0.7.1
requests==2.32.3
urllib3==2.2.2
PyJWT==2.9.0
zaproxy @ git+https://github.com/zaproxy/zap-api-python.git@d826819aa1c96e0071119bef3162cf0c3f183cd2
bcrypt
Flask-Mail
